requirements.txt
numpy
pandas
argparse
pyyaml
pyntcloud
h5py
transforms3d
torchinfo
tensorboard